Linux: Fix compile error
git-svn-id: https://chromiumembedded.googlecode.com/svn/trunk@1436 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ tests/unittests/request_unittest.cc
@@ -240,7 +240,7 @@ class TypeExpectations {
   TypeExpectations(bool browser_side)
       : browser_side_(browser_side) {
     // Build the map of relevant requests.
-    for (int i = 0; i < sizeof(g_type_expected) / sizeof(TypeExpected); ++i) {
+    for (size_t i = 0; i < sizeof(g_type_expected) / sizeof(TypeExpected); ++i) {
       if (g_type_expected[i].browser_side != browser_side_)
         continue;
 
@@ -282,7 +282,7 @@ class TypeExpectations {
 
   // Test if all expectations have been met.
   bool IsDone(bool assert) {
-    for (int i = 0; i < sizeof(g_type_expected) / sizeof(TypeExpected); ++i) {
+    for (size_t i = 0; i < sizeof(g_type_expected) / sizeof(TypeExpected); ++i) {
       if (g_type_expected[i].browser_side != browser_side_)
         continue;
 
@@ -307,7 +307,7 @@ class TypeExpectations {
   int GetExpectedIndex(const std::string& file,
                        cef_transition_type_t transition_type,
                        cef_resource_type_t resource_type) {
-    for (int i = 0; i < sizeof(g_type_expected) / sizeof(TypeExpected); ++i) {
+    for (size_t i = 0; i < sizeof(g_type_expected) / sizeof(TypeExpected); ++i) {
       if (g_type_expected[i].file == file &&
           g_type_expected[i].browser_side == browser_side_ &&
           g_type_expected[i].transition_type == transition_type &&
